lesson 6-4 v2

+++ 06_InheritanceAndPolymorphism/Company/src/company/Employee.java
@@ -2,6 +2,6 @@ package company;
 
 
 public interface Employee {
-      double getMonthSalary();
+      int getMonthSalary();
 
 }
